# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/users": {
      "post": {
        "operationId": "create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      },
      "get": {
        "operationId": "findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/api/v1/users/{id}": {
      "get": {
        "operationId": "findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      },
      "patch": {
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      },
      "delete": {
        "operationId": "remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "operationId": "signUp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "signIn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/refresh-token": {
      "post": {
        "operationId": "refreshToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    }
  },
  "info": {
    "title": "default API",
    "description": "API Documentation",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "description": "User's first name",
            "example": "John",
            "minLength": 2,
            "maxLength": 50
          },
          "lastName": {
            "type": "string",
            "description": "User's last name",
            "example": "Doe",
            "maxLength": 50
          },
          "birthday": {
            "type": "string",
            "description": "User's date of birth (ISO format: YYYY-MM-DD)",
            "example": "1990-02-15"
          },
          "phone": {
            "type": "string",
            "description": "User's phone number (Uzbekistan format: +998XXXXXXXXX)",
            "example": "+998901234567",
            "maxLength": 13,
            "minLength": 13
          },
          "password": {
            "type": "string",
            "description": "User's password (min 6 characters, must include letters and numbers)",
            "example": "Pass123",
            "minLength": 6
          },
          "role": {
            "type": "string",
            "description": "User role",
            "enum": [
              "admin",
              "user"
            ],
            "example": "user"
          }
        },
        "required": [
          "firstName",
          "birthday",
          "phone",
          "password"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "description": "User's first name",
            "example": "John",
            "minLength": 2,
            "maxLength": 50
          },
          "lastName": {
            "type": "string",
            "description": "User's last name",
            "example": "Doe",
            "maxLength": 50
          },
          "birthday": {
            "type": "string",
            "description": "User's date of birth (ISO format: YYYY-MM-DD)",
            "example": "1990-02-15"
          },
          "phone": {
            "type": "string",
            "description": "User's phone number (Uzbekistan format: +998XXXXXXXXX)",
            "example": "+998901234567",
            "maxLength": 13,
            "minLength": 13
          },
          "password": {
            "type": "string",
            "description": "User's password (min 6 characters, must include letters and numbers)",
            "example": "Pass123",
            "minLength": 6
          }
        }
      },
      "RegisterDto": {
        "type": "object",
        "properties": {}
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "description": "User's phone number (Uzbekistan format: +998XXXXXXXXX)",
            "example": "+998901234567",
            "maxLength": 13,
            "minLength": 13
          },
          "password": {
            "type": "string",
            "description": "User's password (min 6 characters, must include letters and numbers)",
            "example": "admin123",
            "minLength": 6
          }
        },
        "required": [
          "phone",
          "password"
        ]
      }
    }
  }
}